fix(config): split pathmanager tests by build tag to fix cross-platform CI
Move linux-specific tests (those referencing linuxPathManager) into
pathmanager_linux_test.go with a //go:build linux constraint. The
generic singleton tests remain in pathmanager_test.go and use
NewPathManager() instead of the linux-only helper.

Fixes: pathmanager_test.go:14:9: undefined: linuxPathManager on
macOS and Windows CI runners.
